services/gateway/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.6
httpx>=0.26
redis>=5.0
prometheus-fastapi-instrumentator>=6.1
